fix two vendors

+++ src/pages/api/profile/delete-member.ts
@@ -3,22 +3,16 @@ import type { APIRoute } from "astro";
 import { createServerClient, parseCookieHeader } from "@supabase/ssr";
 
 export const DELETE: APIRoute = async ({ request, cookies }) => {
-  const supabase = createServerClient(
-    import.meta.env.SUPABASE_URL,
-    import.meta.env.SUPABASE_KEY,
-    {
-      cookies: {
-        getAll() {
-          return parseCookieHeader(request.headers.get("Cookie") ?? "");
-        },
-        setAll(cookiesToSet) {
-          cookiesToSet.forEach(({ name, value, options }) =>
-            cookies.set(name, value, options)
-          );
-        },
+  const supabase = createServerClient(import.meta.env.SUPABASE_URL, import.meta.env.SUPABASE_KEY, {
+    cookies: {
+      getAll() {
+        return parseCookieHeader(request.headers.get("Cookie") ?? "");
       },
-    }
-  );
+      setAll(cookiesToSet) {
+        cookiesToSet.forEach(({ name, value, options }) => cookies.set(name, value, options));
+      },
+    },
+  });
 
   const { id } = await request.json();
   const userId = (await supabase.auth.getUser()).data.user?.id;
@@ -29,14 +23,9 @@ export const DELETE: APIRoute = async ({ request, cookies }) => {
 
   await supabase.from("transactions").delete().eq("profile_id", id);
 
-  const { error } = await supabase
-    .from("profiles")
-    .delete()
-    .eq("id", id)
-    .eq("parent_id", userId);
+  const { error } = await supabase.from("profiles").delete().eq("id", id).eq("parent_id", userId);
 
   if (error) {
-    console.log(error);
     return new Response(`Error deleting family member: ${error.message}`, {
       status: 500,
     });
